refactor: split DiffContent into dedicated section components (#111)
Extract 7 sub-components from diff-content.tsx (~470 lines → ~90 lines):
- diff-header.tsx: Badge + comparison links + datetime
- diff-status-message.tsx: status-based messages (no_baseline/unavailable/failed)
- file-changes-section.tsx: FileChangesSection
- bom-changes-section.tsx: BomChangesSection
- checks-section.tsx: ChecksSection
- artifact-changes-section.tsx: ArtifactChangesSection
- preview-links-section.tsx: PreviewLinksSection

Pure code move, no behavior changes.
